Use MessagePack pack() instead of Python representation repr().
This patch changes the use of `repr()` to the MessagePack function `pack()` so that languages other than Python can access the data that  is recorded. The name of the streams are renamed to `Python MessagePack Data - Eye 0/1` and `Gaze MessagePack Data`. The names of some functions are also changed to reflect the new method.